Finishing touches on EstimateDialog extraction refactor.
